API Documentation

API Documentation

Documentation for the POMDPs.jl user interface. You can get help for any type or function in the module by typing ? in the Julia REPL followed by the name of type or function. For example:

julia> using POMDPs
julia> ?
help?> reward
search: reward

  reward{S,A,O}(pomdp::POMDP{S,A,O}, state::S, action::A, statep::S)

  Returns the immediate reward for the s-a-s triple

  reward{S,A,O}(pomdp::POMDP{S,A,O}, state::S, action::A)

  Returns the immediate reward for the s-a pair

Contents

API Documentation

Contents

Index

Types

Model Functions

Distribution/Space Functions

Belief Functions

Policy and Solver Functions

Simulator

Other

Index

POMDPs.MDP

POMDPs.POMDP

POMDPs.Policy

POMDPs.Simulator

POMDPs.Solver

POMDPs.Updater

Base.rand

Distributions.pdf

Distributions.support

POMDPs.action

POMDPs.actionindex

POMDPs.actions

POMDPs.actiontype

POMDPs.add_registry

POMDPs.available

POMDPs.check_requirements

POMDPs.convert_a

POMDPs.convert_o

POMDPs.convert_s

POMDPs.dimensions

POMDPs.discount

POMDPs.generate_o

POMDPs.generate_or

POMDPs.generate_s

POMDPs.generate_so

POMDPs.generate_sor

POMDPs.generate_sr

POMDPs.get_requirements

POMDPs.implemented

POMDPs.initialize_belief

POMDPs.initialstate

POMDPs.initialstate_distribution

POMDPs.isterminal

POMDPs.n_actions

POMDPs.n_observations

POMDPs.n_states

POMDPs.observation

POMDPs.observations

POMDPs.obsindex

POMDPs.obstype

POMDPs.requirements_info

POMDPs.reward

POMDPs.sampletype

POMDPs.show_requirements

POMDPs.simulate

POMDPs.solve

POMDPs.stateindex

POMDPs.states

POMDPs.statetype

POMDPs.transition

POMDPs.update

POMDPs.updater

POMDPs.value

Statistics.mean

StatsBase.mode

POMDPs.@POMDP_require

POMDPs.@POMDP_requirements

POMDPs.@get_requirements

POMDPs.@req

POMDPs.@requirements_info

POMDPs.@show_requirements

POMDPs.@subreq

POMDPs.@warn_requirements

Types

POMDPs.POMDP — Type.

POMDP{S,A,O}

Abstract base type for a partially observable Markov decision process.

S: state type
A: action type
O: observation type

source

POMDPs.MDP — Type.

MDP{S,A}

Abstract base type for a fully observable Markov decision process.

S: state type
A: action type

source

POMDPs.Solver — Type.

Base type for an MDP/POMDP solver

source

POMDPs.Policy — Type.

Base type for a policy (a map from every possible belief, or more abstract policy state, to an optimal or suboptimal action)

source

POMDPs.Updater — Type.

Abstract type for an object that defines how the belief should be updated

A belief is a general construct that represents the knowledge an agent has about the state of the system. This can be a probability distribution, an action observation history or a more general representation.

source

Model Functions

Explicit

These functions return distributions.

POMDPs.transition — Function.

transition(problem::POMDP, state, action)
transition(problem::MDP, state, action)

Return the transition distribution from the current state-action pair

source

POMDPs.observation — Function.

observation(problem::POMDP, statep)
observation(problem::POMDP, action, statep)
observation(problem::POMDP, state, action, statep)

Return the observation distribution. You need only define the method with the fewest arguments needed to determine the observation distribution.

Example

using POMDPModelTools # for SparseCat

struct MyPOMDP <: POMDP{Int, Int, Int} end

observation(p::MyPOMDP, sp::Int) = SparseCat([sp-1, sp, sp+1], [0.1, 0.8, 0.1])

source

POMDPs.initialstate_distribution — Function.

initialstate_distribution(pomdp::POMDP)
initialstate_distribution(mdp::MDP)

Return a distribution of the initial state of the pomdp or mdp.

source

Generative

These functions should return states, observations, and rewards.

POMDPs.generate_s — Function.

generate_s{S,A}(p::Union{POMDP{S,A},MDP{S,A}}, s::S, a::A, rng::AbstractRNG)

Return the next state given current state s and action taken a.

source

POMDPs.generate_o — Function.

generate_o{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, sp::S, rng::AbstractRNG)

Return the next observation given current state s, action taken a and next state sp.

Usually the observation would only depend on the next state sp.

generate_o{S,A,O}(p::POMDP{S,A,O}, s::S, rng::AbstractRNG)

Return the observation from the current state. This should be used to generate initial observations.

source

POMDPs.generate_sr — Function.

generate_sr{S}(p::Union{POMDP{S},MDP{S}}, s, a, rng::AbstractRNG)

Return the next state sp and reward for taking action a in current state s.

source

POMDPs.generate_so — Function.

generate_so{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, rng::AbstractRNG)

Return the next state sp and observation o.

source

POMDPs.generate_or — Function.

generate_or{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, sp::S, rng::AbstractRNG)

Return the observation o and reward for taking action a in current state s reaching state sp.

source

POMDPs.generate_sor — Function.

generate_sor{S,A,O}(p::POMDP{S,A,O}, s::S, a::A, rng::AbstractRNG)

Return the next state sp, observation o and reward for taking action a in current state s.

source

POMDPs.initialstate — Function.

initialstate{S}(p::Union{POMDP{S},MDP{S}}, rng::AbstractRNG)

Return the initial state for the problem p.

Usually the initial state is sampled from an initial state distribution.

source

Common

POMDPs.states — Function.

states(problem::POMDP)
states(problem::MDP)

Returns the complete state space of a POMDP.

source

POMDPs.actions — Function.

actions(problem::POMDP)
actions(problem::MDP)

Returns the entire action space of a POMDP.

source

POMDPs.observations — Function.

observations(problem::POMDP)

Return the entire observation space.

source

POMDPs.reward — Function.

reward(m::POMDP, s, a)
reward(m::MDP, s, a)

Return the immediate reward for the s-a pair.

For some problems, it is easier to express reward(m, s, a, sp) than reward(m, s, a), but some solvers, e.g. SARSOP, can only use reward(m, s, a). Both can be implemented for a problem, but when reward(m, s, a) is implemented, it should be consistent with reward(m, s, a, sp), that is, it should be the expected value over all destination states.

source

POMDPs.isterminal — Function.

isterminal(m::Union{MDP,POMDP}, s)

Check if state s is terminal.

If a state is terminal, no actions will be taken in it and no additional rewards will be accumulated. Thus, the value at such a state is, by definition, zero.

source

POMDPs.discount — Function.

discount(problem::POMDP)
discount(problem::MDP)

Return the discount factor for the problem.

source

POMDPs.n_states — Function.

n_states(problem::POMDP)
n_states(problem::MDP)

Return the number of states in problem. Used for discrete models only.

source

POMDPs.n_actions — Function.

n_actions(problem::POMDP)
n_actions(problem::MDP)

Return the number of actions in problem. Used for discrete models only.

source

POMDPs.n_observations — Function.

n_observations(problem::POMDP)

Return the number of observations in problem. Used for discrete models only.

source

POMDPs.stateindex — Function.

stateindex(problem::POMDP, s)
stateindex(problem::MDP, s)

Return the integer index of state s. Used for discrete models only.

source

POMDPs.actionindex — Function.

actionindex(problem::POMDP, a)
actionindex(problem::MDP, a)

Return the integer index of action a. Used for discrete models only.

source

POMDPs.obsindex — Function.

obsindex(problem::POMDP, o)

Return the integer index of observation o. Used for discrete models only.

source

POMDPs.convert_s — Function.

convert_s(::Type{V}, s, problem::Union{MDP,POMDP}) where V<:AbstractArray
convert_s(::Type{S}, vec::V, problem::Union{MDP,POMDP}) where {S,V<:AbstractArray}

Convert a state to vectorized form or vice versa.

source

POMDPs.convert_a — Function.

convert_a(::Type{V}, a, problem::Union{MDP,POMDP}) where V<:AbstractArray
convert_a(::Type{A}, vec::V, problem::Union{MDP,POMDP}) where {A,V<:AbstractArray}

Convert an action to vectorized form or vice versa.

source

POMDPs.convert_o — Function.

convert_o(::Type{V}, o, problem::Union{MDP,POMDP}) where V<:AbstractArray
convert_o(::Type{O}, vec::V, problem::Union{MDP,POMDP}) where {O,V<:AbstractArray}

Convert an observation to vectorized form or vice versa.

source

Distribution/Space Functions

Base.rand — Function.

rand(rng::AbstractRNG, d::Any)

Return a random element from distribution or space d.

If d is a state or transition distribution, the sample will be a state; if d is an action distribution, the sample will be an action or if d is an observation distribution, the sample will be an observation.

source

Distributions.pdf — Function.

pdf(d::Any, x::Any)

Evaluate the probability density of distribution d at sample x.

source

StatsBase.mode — Function.

mode(d::Any)

Return the most likely value in a distribution d.

source

Statistics.mean — Function.

mean(d::Any)

Return the mean of a distribution d.

source

POMDPs.dimensions — Function.

dimensions(s::Any)

Returns the number of dimensions in space s.

source

Distributions.support — Function.

support(d::Any)

Return an iterable object containing the possible values that can be sampled from distribution d. Values with zero probability may be skipped.

source

POMDPs.sampletype — Function.

sampletype(T::Type)
sampletype(d::Any) = sampletype(typeof(d))

Return the type of objects that are sampled from a distribution or space d when rand(rng, d) is called.

The distribution writer should implement the sampletype(::Type) method for the distribution type, then the function can be called for that type or for objects of that type (i.e. the sampletype(d::Any) = sampletype(typeof(d)) default is provided).

source

Belief Functions

POMDPs.update — Function.

update(updater::Updater, belief_old, action, observation)

Return a new instance of an updated belief given belief_old and the latest action and observation.

source

POMDPs.initialize_belief — Function.

initialize_belief(updater::Updater,
                     state_distribution::Any)
initialize_belief(updater::Updater, belief::Any)

Returns a belief that can be updated using updater that has similar distribution to state_distribution or belief.

The conversion may be lossy. This function is also idempotent, i.e. there is a default implementation that passes the belief through when it is already the correct type: initialize_belief(updater::Updater, belief) = belief

source

Policy and Solver Functions

POMDPs.solve — Function.

solve(solver::Solver, problem::POMDP)

Solves the POMDP using method associated with solver, and returns a policy.

source

POMDPs.updater — Function.

updater(policy::Policy)

Returns a default Updater appropriate for a belief type that policy p can use

source

POMDPs.action — Function.

action(policy::Policy, x)

Returns the action that the policy deems best for the current state or belief, x.

x is a generalized information state - can be a state in an MDP, a distribution in POMDP, or another specialized policy-dependent representation of the information needed to choose an action.

source

POMDPs.value — Function.

value(p::Policy, s)
value(p::Policy, s, a)

Returns the utility value from policy p given the state (or belief), or state-action (or belief-action) pair.

The state-action version is commonly referred to as the Q-value.

source

Simulator

POMDPs.Simulator — Type.

Base type for an object defining how simulations should be carried out.

source

POMDPs.simulate — Function.

simulate(sim::Simulator, m::POMDP, p::Policy, u::Updater=updater(p), b0=initialstate_distribution(m), s0=initialstate(m, rng))
simulate(sim::Simulator, m::MDP, p::Policy, s0=initialstate(m, rng))

Run a simulation using the specified policy.

The return type is flexible and depends on the simulator. Simulations should adhere to the Simulation Standard.

source

Other

The following functions are not part of the API for specifying and solving POMDPs, but are included in the package.

Type Inference

POMDPs.statetype — Function.

statetype(t::Type)
statetype(p::Union{POMDP,MDP})

Return the state type for a problem type (the S in POMDP{S,A,O}).

type A <: POMDP{Int, Bool, Bool} end

statetype(A) # returns Int

source

POMDPs.actiontype — Function.

actiontype(t::Type)
actiontype(p::Union{POMDP,MDP})

Return the state type for a problem type (the S in POMDP{S,A,O}).

type A <: POMDP{Bool, Int, Bool} end

actiontype(A) # returns Int

source

POMDPs.obstype — Function.

obstype(t::Type)

Return the state type for a problem type (the S in POMDP{S,A,O}).

type A <: POMDP{Bool, Bool, Int} end

obstype(A) # returns Int

source

Requirements Specification

POMDPs.check_requirements — Function.

check_requirements(r::AbstractRequirementSet)

Check whether the methods in r have implementations with implemented(). Return true if all methods have implementations.

source

POMDPs.show_requirements — Function.

show_requirements(r::AbstractRequirementSet)

Check whether the methods in r have implementations with implemented() and print out a formatted list showing which are missing. Return true if all methods have implementations.

source

POMDPs.get_requirements — Function.

get_requirements(f::Function, args::Tuple)

Return a RequirementSet for the function f and arguments args.

source

POMDPs.requirements_info — Function.

requirements_info(s::Solver, p::Union{POMDP,MDP}, ...)

Print information about the requirement for solver s.

source

POMDPs.@POMDP_require — Macro.

@POMDP_require solve(s::CoolSolver, p::POMDP) begin
    PType = typeof(p)
    @req states(::PType)
    @req actions(::PType)
    @req transition(::PType, ::S, ::A)
    s = first(states(p))
    a = first(actions(p))
    t_dist = transition(p, s, a)
    @req rand(::AbstractRNG, ::typeof(t_dist))
end

Create a get_requirements implementation for the function signature and the requirements block.

source

POMDPs.@POMDP_requirements — Macro.

reqs = @POMDP_requirements CoolSolver begin
    PType = typeof(p)
    @req states(::PType)
    @req actions(::PType)
    @req transition(::PType, ::S, ::A)
    s = first(states(p))
    a = first(actions(p))
    t_dist = transition(p, s, a)
    @req rand(::AbstractRNG, ::typeof(t_dist))
end

Create a RequirementSet object.

source

POMDPs.@requirements_info — Macro.

@requirements_info ASolver() [YourPOMDP()]

Print information about the requirements for a solver.

source

POMDPs.@get_requirements — Macro.

@get_requirements f(arg1, arg2)

Call get_requirements(f, (arg1,arg2)).

source

POMDPs.@show_requirements — Macro.

@show_requirements solve(solver, problem)

Print a a list of requirements for a function call.

source

POMDPs.@warn_requirements — Macro.

@warn_requirements solve(solver, problem)

Print a warning if there are unmet requirements.

source

POMDPs.@req — Macro.

@req f( ::T1, ::T2)

Convert a f( ::T1, ::T2) expression to a (f, Tuple{T1,T2})::Req for pushing to a RequirementSet.

If in a @POMDP_requirements or @POMDP_require block, marks the requirement for including in the set of requirements.

source

POMDPs.@subreq — Macro.

@subreq f(arg1, arg2)

In a @POMDP_requirements or @POMDP_require block, include the requirements for f(arg1, arg2) as a child argument set.

source

POMDPs.implemented — Function.

implemented(function, Tuple{Arg1Type, Arg2Type})

Check whether there is an implementation available that will return a suitable value.

source

Utility Tools

POMDPs.add_registry — Function.

add_registry()

Adds the JuliaPOMDP registry

source

POMDPs.available — Function.

available()

Prints all the available packages in the JuliaPOMDP registry

source